p92335: long 으로 타입 변환 / 88.1 점 시간초과로 실패

+++ src/main/java/org/example/solution/programers/lv2/p92335/Solution_92335.java
@@ -15,7 +15,7 @@ public class Solution_92335 {
         int answer = 0;
         for (int i = 0; i < split.length; i++) {
             if (split[i].equals("")) continue;
-            int num = Integer.parseInt(split[i]);
+            long num = Long.parseLong(split[i]);
             boolean check = false;
 
             for (int j = 2; j < num; j++) {
